Implement q_delete_dup function
Initialize flag variable as false and use the list_for_each_entry_safe
function from list.h to through each node. Check if the current node's
value is equal to the next node's value. If they are equal, free the
current node and set the flag to true. Even if the last duplicate node
is not equal to the next node, it will still be deleted because the
flag is true.