Implement q_shuffle function
Use the q_size function to get the length of the queue. Randomly pick a
number random from 0 to (len - 1). Then, old will point to the random-th
node counting from the front, and new will point to the last node that
has not been picked. Swap the values of the nodes pointed to by old and
new, and then decrement the length (len) by 1. Repeat this process until
all nodes have been picked, and the shuffle is complete.